requirements.txt
BeautifulSoup==3.2.1
beautifulsoup4==4.4.1
cssselect==0.9.1
goose-extractor==1.0.25
jieba==0.38
langid==1.1.5
lxml==3.6.0
nltk==3.2
numpy==1.10.4
Pillow==3.1.1
pyfscache==0.9.12
requests==2.9.1
scikit-learn==0.17.1
scipy==0.17.0
textblob==0.11.1
treelib==1.3.2
